+++ packages/block-library/src/accordion-item/edit.js
@@ -56,7 +56,6 @@ export default function Edit( {
 
 	const innerBlocksProps = useInnerBlocksProps( blockProps, {
 		templateLock: 'all',
-		directInsert: true,
 	} );
 
 	return (