+++ packages/block-library/src/accordion/index.js
@@ -52,6 +52,7 @@ export const settings = {
 	},
 	template: TEMPLATE,
 	templateInsertUpdatesSelection: true,
+	defaultBlock: { name: 'core/accordion-item' },
 	edit,
 	save,
 };