+++ packages/block-library/src/navigation/edit/unsaved-inner-blocks.js
@@ -11,7 +11,7 @@ import { useContext, useEffect, useRef } from '@wordpress/element';
  * Internal dependencies
  */
 import { areBlocksDirty } from './are-blocks-dirty';
-import { DEFAULT_BLOCK, SELECT_NAVIGATION_MENUS_ARGS } from '../constants';
+import { SELECT_NAVIGATION_MENUS_ARGS } from '../constants';
 
 const EMPTY_OBJECT = {};
 
@@ -52,7 +52,6 @@ export default function UnsavedInnerBlocks( {
 		},
 		{
 			renderAppender: hasSelection ? undefined : false,
-			defaultBlock: DEFAULT_BLOCK,
 			directInsert: true,
 		}
 	);